Moved files for atributes around
