+++ react/components/types.ts
@@ -13,3 +13,18 @@ export interface AutoCompleteProps {
   closeMenu: () => void
   isOpen: boolean
 }
+
+export interface Item {
+  label: string | JSX.Element
+  value: string
+  link: string
+  attributes?: AttributeItem[]
+}
+
+export interface AttributeItem {
+  groupValue: string
+  value: string
+  label: string
+  link: string
+  key: string
+}
